[ fix ] Footnotes can contain multiple paragraphs
Patched parser / pretty-printer to be able to include arbitrary
"Structures" in the footnotes. The main reason is to be able to
include code blocks.

I used this opportunity to clean up the way the list of footnotes
is displayed. Each footnote is a p block in which we have  a left
floating div with the number / link to jump back to where it was
referenced and another div containing the body of the footnote.

+-----+---------------------------+
| [1] | The body of the footnote  |
|     | goes in this div          |
+-----+---------------------------+
+-----+---------------------------+
| [2] | And the next footnote sits|
|     | under it.                 |
|     | And so on.                |
+-----+---------------------------+